+++ src/schematic/renderer/index.ts
@@ -1,3 +1,2 @@
 export * from './chained'
-export * from './util'
 export * from './background'
