+++ src/util/graphics/index.ts
@@ -1,6 +1,6 @@
 export * from './tint_image'
 export * from './block_asset'
-export * from './translate_pos'
+export * from './positioning'
 export * from './outline_image'
 export * from './draw_rotated'
 export * from './draw_bridge'
